litt opprydding i api kall
* flytt urler fra lenker.ts til der de er brukt
* legg til zod og sentry og bedre feilmelding der det mangler
