+++ src/api/altinnApi.ts
@@ -1,6 +1,11 @@
-import { altinnUrl } from '../lenker';
 import { gittMiljo } from '../utils/environment';
 import { navtjenester } from '../altinn/tjenester';
+
+export const altinnUrl = gittMiljo({
+    prod: 'https://altinn.no',
+    dev: 'https://tt02.altinn.no',
+    other: '/min-side-arbeidsgiver/mock/tt02.altinn.no',
+});
 
 export enum Status {
     Ulest = 'Ulest',
