+++ src/api/pamApi.ts
@@ -1,5 +1,14 @@
-import { pamSettBedriftURL, pamHentStillingsannonserURL } from '../lenker';
+import {gittMiljo} from "../utils/environment";
 
+
+const pamSettBedriftURL = gittMiljo({
+    prod: (orgnr: string) =>
+        `https://arbeidsplassen.nav.no/stillingsregistrering-api/api/arbeidsgiver/${orgnr}`,
+    dev: (orgnr: string) =>
+        `https://arbeidsplassen.dev.nav.no/stillingsregistrering-api/api/arbeidsgiver/${orgnr}`,
+    other: (orgnr: string) =>
+        `/min-side-arbeidsgiver/mock/arbeidsplassen.nav.no/stillingsregistrering-api/api/arbeidsgiver/${orgnr}`,
+});
 export const settBedriftIPam = (orgnr: string): Promise<unknown> =>
     fetch(pamSettBedriftURL(orgnr), {
         method: 'GET',
@@ -17,6 +26,11 @@ interface PamStatusAnnonser {
     PUBLISERT: number;
 }
 
+const pamHentStillingsannonserURL = gittMiljo({
+    prod: 'https://arbeidsplassen.nav.no/stillingsregistrering-api/api/stillinger/numberByStatus',
+    dev: 'https://arbeidsplassen.dev.nav.no/stillingsregistrering-api/api/stillinger/numberByStatus',
+    other: '/min-side-arbeidsgiver/mock/arbeidsplassen.nav.no/stillingsregistrering-api/api/stillinger/numberByStatus',
+});
 export const hentAntallannonser = async (): Promise<number> => {
     const respons = await fetch(pamHentStillingsannonserURL, {
         method: 'GET',
